Fix adding newline error
`T.unlines()` in the function `saveNote` add a newline
at the end of the returned string:

    > T.unlines ["Title", "body"]
    "Title\nbody\n"

Every time a note edited, a newline is appended to the end of the note.
So we use `T.strip()` to get rid of it.
The problem is the ending newline of the note itself will be removed, too:

    > T.strip $ T.unlines ["Title", "body\n"]
    Title\nbody"